3: Renamed randomvariablelist.py to private
