+++ src/probnum/filtsmooth/gaussfiltsmooth/kalmanposterior.py
@@ -5,7 +5,7 @@ Contains the discrete time and function outputs.
 Provides dense output by being callable.
 Can function values can also be accessed by indexing.
 """
-from probnum.prob.randomvariablelist import _RandomVariableList
+from probnum.prob._randomvariablelist import _RandomVariableList
 from probnum.filtsmooth.filtsmoothposterior import FiltSmoothPosterior
